Introduced a UserSessionFlowCoordinator and split up the AppCoordinator state machine
