+++ ElementX/Sources/Application/AppCoordinator.swift
@@ -31,7 +31,7 @@ struct ServiceLocator {
     let userIndicatorPresenter: UserIndicatorTypePresenter
 }
 
-class AppCoordinator: Coordinator {
+class AppCoordinator: Coordinator, UserSessionFlowCoordinatorDelegate {
     private let window: UIWindow
     
     private let stateMachine: AppCoordinatorStateMachine
@@ -52,8 +52,9 @@ class AppCoordinator: Coordinator {
         }
     }
     
+    private var userSessionFlowCoordinator: UserSessionFlowCoordinator?
+    
     private let bugReportService: BugReportServiceProtocol
-    private let screenshotDetector: ScreenshotDetector
     private let backgroundTaskService: BackgroundTaskServiceProtocol
 
     private var loadingIndicator: UserIndicator?
@@ -89,9 +90,6 @@ class AppCoordinator: Coordinator {
         
         userSessionStore = UserSessionStore(bundleIdentifier: bundleIdentifier,
                                             backgroundTaskService: backgroundTaskService)
-
-        screenshotDetector = ScreenshotDetector()
-        screenshotDetector.callback = processScreenshotDetection
         
         setupStateMachine()
         
@@ -107,6 +105,12 @@ class AppCoordinator: Coordinator {
 
     func stop() {
         hideLoadingIndicator()
+    }
+    
+    // MARK: - UserSessionFlowCoordinatorDelegate
+    
+    func userSessionFlowCoordinatorDidRequestSignOut(_ userSessionFlowCoordinator: UserSessionFlowCoordinator) {
+        stateMachine.processEvent(.signOut)
     }
 
     // MARK: - Private
@@ -142,43 +146,29 @@ class AppCoordinator: Coordinator {
             switch (context.fromState, context.event, context.toState) {
             case (.initial, .startWithAuthentication, .signedOut):
                 self.startAuthentication()
-            case (.signedOut, .succeededSigningIn, .homeScreen):
-                self.presentHomeScreen()
-                
+            case (.signedOut, .succeededSigningIn, .signedIn):
+                self.setupUserSession()
             case (.initial, .startWithExistingSession, .restoringSession):
                 self.showLoadingIndicator()
                 self.restoreUserSession()
             case (.restoringSession, .failedRestoringSession, .signedOut):
                 self.hideLoadingIndicator()
                 self.showLoginErrorToast()
-            case (.restoringSession, .succeededRestoringSession, .homeScreen):
+            case (.restoringSession, .succeededRestoringSession, .signedIn):
                 self.hideLoadingIndicator()
-                self.presentHomeScreen()
-            
-            case(_, _, .roomScreen(let roomId)):
-                self.presentRoomWithIdentifier(roomId)
-            case(.roomScreen(let roomId), .dismissedRoomScreen, .homeScreen):
-                self.tearDownDismissedRoomScreen(roomId)
-
+                self.setupUserSession()
             case (_, .signOut, .signingOut):
                 self.showLoadingIndicator()
                 self.tearDownUserSession()
             case (.signingOut, .completedSigningOut, .signedOut):
                 self.presentSplashScreen()
                 self.hideLoadingIndicator()
-
             case (_, .remoteSignOut(let isSoft), .remoteSigningOut):
                 self.showLoadingIndicator()
                 self.tearDownUserSession(isSoftLogout: isSoft)
             case (.remoteSigningOut(let isSoft), .completedSigningOut, .signedOut):
                 self.presentSplashScreen(isSoftLogout: isSoft)
                 self.hideLoadingIndicator()
-
-            case (.homeScreen, .showSessionVerificationScreen, .sessionVerificationScreen):
-                self.presentSessionVerification()
-            case (.sessionVerificationScreen, .dismissedSessionVerificationScreen, .homeScreen):
-                self.tearDownDismissedSessionVerificationScreen()
-                
             default:
                 fatalError("Unknown transition: \(context)")
             }
@@ -257,8 +247,21 @@ class AppCoordinator: Coordinator {
         }
     }
     
+    private func setupUserSession() {
+        let userSessionFlowCoordinator = UserSessionFlowCoordinator(userSession: userSession,
+                                                                    navigationRouter: navigationRouter,
+                                                                    bugReportService: bugReportService)
+        userSessionFlowCoordinator.delegate = self
+        userSessionFlowCoordinator.start()
+        
+        navigationRouter.setRootModule(userSessionFlowCoordinator)
+        
+        self.userSessionFlowCoordinator = userSessionFlowCoordinator
+    }
+    
     private func tearDownUserSession(isSoftLogout: Bool = false) {
         userSession.clientProxy.stopSync()
+        userSessionFlowCoordinator?.stop()
         
         deobserveUserSessionChanges()
         
@@ -291,38 +294,6 @@ class AppCoordinator: Coordinator {
         }
     }
     
-    private func presentHomeScreen() {
-        userSession.clientProxy.startSync()
-        
-        let parameters = HomeScreenCoordinatorParameters(userSession: userSession,
-                                                         attributedStringBuilder: AttributedStringBuilder())
-        let coordinator = HomeScreenCoordinator(parameters: parameters)
-        
-        coordinator.callback = { [weak self] action in
-            guard let self = self else { return }
-            
-            switch action {
-            case .presentRoom(let roomIdentifier):
-                self.stateMachine.processEvent(.showRoomScreen(roomId: roomIdentifier))
-            case .presentSettings:
-                self.presentSettingsScreen()
-            case .presentBugReport:
-                self.presentBugReportScreen()
-            case .verifySession:
-                self.stateMachine.processEvent(.showSessionVerificationScreen)
-            case .signOut:
-                self.stateMachine.processEvent(.signOut)
-            }
-        }
-        
-        add(childCoordinator: coordinator)
-        navigationRouter.setRootModule(coordinator)
-
-        if bugReportService.crashedLastRun {
-            showCrashPopup()
-        }
-    }
-
     private func observeUserSessionChanges() {
         userSession.callbacks
             .receive(on: DispatchQueue.main)
@@ -345,179 +316,6 @@ class AppCoordinator: Coordinator {
     private func deobserveUserSessionChanges() {
         cancellables.forEach { $0.cancel() }
         cancellables.removeAll()
-    }
-    
-    // MARK: Rooms
-
-    private func presentRoomWithIdentifier(_ roomIdentifier: String) {
-        guard let roomProxy = userSession.clientProxy.roomForIdentifier(roomIdentifier) else {
-            MXLog.error("Invalid room identifier: \(roomIdentifier)")
-            return
-        }
-        let userId = userSession.clientProxy.userIdentifier
-        
-        let timelineItemFactory = RoomTimelineItemFactory(userID: userId,
-                                                          mediaProvider: userSession.mediaProvider,
-                                                          roomProxy: roomProxy,
-                                                          attributedStringBuilder: AttributedStringBuilder())
-
-        let timelineController = RoomTimelineController(userId: userId,
-                                                        roomId: roomIdentifier,
-                                                        timelineProvider: roomProxy.timelineProvider,
-                                                        timelineItemFactory: timelineItemFactory,
-                                                        mediaProvider: userSession.mediaProvider,
-                                                        roomProxy: roomProxy)
-
-        let parameters = RoomScreenCoordinatorParameters(timelineController: timelineController,
-                                                         mediaProvider: userSession.mediaProvider,
-                                                         roomName: roomProxy.displayName ?? roomProxy.name,
-                                                         roomAvatarUrl: roomProxy.avatarURL)
-        let coordinator = RoomScreenCoordinator(parameters: parameters)
-
-        add(childCoordinator: coordinator)
-        navigationRouter.push(coordinator) { [weak self] in
-            guard let self = self else { return }
-            self.stateMachine.processEvent(.dismissedRoomScreen)
-        }
-    }
-    
-    private func tearDownDismissedRoomScreen(_ roomId: String) {
-        guard let coordinator = childCoordinators.last as? RoomScreenCoordinator else {
-            fatalError("Invalid coordinator hierarchy: \(childCoordinators)")
-        }
-        
-        remove(childCoordinator: coordinator)
-    }
-    
-    // MARK: Settings
-    
-    private func presentSettingsScreen() {
-        let navController = ElementNavigationController()
-        let newNavigationRouter = NavigationRouter(navigationController: navController)
-
-        let parameters = SettingsCoordinatorParameters(navigationRouter: newNavigationRouter,
-                                                       userSession: userSession,
-                                                       bugReportService: bugReportService)
-        let coordinator = SettingsCoordinator(parameters: parameters)
-        coordinator.callback = { [weak self] action in
-            guard let self = self else { return }
-            switch action {
-            case .dismiss:
-                self.dismissSettingsScreen()
-            case .logout:
-                self.dismissSettingsScreen()
-                self.stateMachine.processEvent(.signOut)
-            }
-        }
-
-        add(childCoordinator: coordinator)
-        coordinator.start()
-        navController.viewControllers = [coordinator.toPresentable()]
-        navigationRouter.present(navController, animated: true)
-    }
-
-    @objc
-    private func dismissSettingsScreen() {
-        MXLog.debug("dismissSettingsScreen")
-
-        guard let coordinator = childCoordinators.first(where: { $0 is SettingsCoordinator }) else {
-            return
-        }
-
-        navigationRouter.dismissModule()
-        remove(childCoordinator: coordinator)
-    }
-    
-    private func showCrashPopup() {
-        let alert = UIAlertController(title: nil,
-                                      message: ElementL10n.sendBugReportAppCrashed,
-                                      preferredStyle: .alert)
-
-        alert.addAction(UIAlertAction(title: ElementL10n.no, style: .cancel))
-        alert.addAction(UIAlertAction(title: ElementL10n.yes, style: .default) { [weak self] _ in
-            self?.presentBugReportScreen()
-        })
-
-        navigationRouter.present(alert, animated: true)
-    }
-
-    private func processScreenshotDetection(image: UIImage?, error: Error?) {
-        MXLog.debug("Detected screenshot: \(String(describing: image)), error: \(String(describing: error))")
-
-        let alert = UIAlertController(title: ElementL10n.screenshotDetectedTitle,
-                                      message: ElementL10n.screenshotDetectedMessage,
-                                      preferredStyle: .alert)
-
-        alert.addAction(UIAlertAction(title: ElementL10n.no, style: .cancel))
-        alert.addAction(UIAlertAction(title: ElementL10n.yes, style: .default) { [weak self] _ in
-            self?.presentBugReportScreen(for: image)
-        })
-
-        navigationRouter.present(alert, animated: true)
-    }
-
-    private func presentBugReportScreen(for image: UIImage? = nil) {
-        let parameters = BugReportCoordinatorParameters(bugReportService: bugReportService,
-                                                        screenshot: image)
-        let coordinator = BugReportCoordinator(parameters: parameters)
-        coordinator.completion = { [weak self, weak coordinator] in
-            guard let self = self, let coordinator = coordinator else { return }
-            self.navigationRouter.dismissModule(animated: true)
-            self.remove(childCoordinator: coordinator)
-        }
-
-        add(childCoordinator: coordinator)
-        coordinator.start()
-        let navController = ElementNavigationController(rootViewController: coordinator.toPresentable())
-        navController.navigationBar.topItem?.leftBarButtonItem = UIBarButtonItem(barButtonSystemItem: .cancel,
-                                                                                 target: self,
-                                                                                 action: #selector(dismissBugReportScreen))
-        navController.isModalInPresentation = true
-        navigationRouter.present(navController, animated: true)
-    }
-
-    @objc
-    private func dismissBugReportScreen() {
-        MXLog.debug("dismissBugReportScreen")
-
-        guard let bugReportCoordinator = childCoordinators.first(where: { $0 is BugReportCoordinator }) else {
-            return
-        }
-
-        navigationRouter.dismissModule()
-        remove(childCoordinator: bugReportCoordinator)
-    }
-    
-    // MARK: Session verification
-        
-    private func presentSessionVerification() {
-        Task {
-            guard let sessionVerificationController = userSession.sessionVerificationController else {
-                fatalError("The sessionVerificationController should aways be valid at this point")
-            }
-            
-            let parameters = SessionVerificationCoordinatorParameters(sessionVerificationControllerProxy: sessionVerificationController)
-            
-            let coordinator = SessionVerificationCoordinator(parameters: parameters)
-            
-            coordinator.callback = { [weak self] in
-                self?.navigationRouter.dismissModule()
-                self?.stateMachine.processEvent(.dismissedSessionVerificationScreen)
-            }
-            
-            add(childCoordinator: coordinator)
-            navigationRouter.present(coordinator)
-
-            coordinator.start()
-        }
-    }
-    
-    private func tearDownDismissedSessionVerificationScreen() {
-        guard let coordinator = childCoordinators.last as? SessionVerificationCoordinator else {
-            fatalError("Invalid coordinator hierarchy: \(childCoordinators)")
-        }
-
-        remove(childCoordinator: coordinator)
     }
     
     // MARK: Toasts and loading indicators
