+++ UITests/SupportingFiles/target.yml
@@ -78,10 +78,10 @@ targets:
         - "**/__Snapshots__/**"
     - path: ../SupportingFiles
     - path: ../../Tools/Scripts/Templates/SimpleScreenExample/Tests/UI
-    - path: ../../ElementX/Sources/BuildSettings.swift
+    - path: ../../ElementX/Sources/Application/BuildSettings.swift
     - path: ../../ElementX/Sources/Screens/RoomScreen/View/Style/TimelineStyle.swift
     - path: ../../ElementX/Sources/Services/Analytics/AnalyticsConfiguration.swift
-    - path: ../../ElementX/Sources/UITestScreenIdentifier.swift
+    - path: ../../ElementX/Sources/UITests/UITestScreenIdentifier.swift
     - path: ../../ElementX/Sources/Generated/Strings.swift
     - path: ../../ElementX/Sources/Generated/Strings+Untranslated.swift
     - path: ../../ElementX/Sources/Generated/InfoPlist.swift
